+++ ChangeLog.md
@@ -1,10 +1,23 @@
 # Release Notes
 KsWare.Presentation.Converters
 
+#### 2020-01-28
+- KsWare.Presentation.Interfaces 0.1.40
+- ResourceConverter
+
+#### 
+- BooleanComparingConverter
+- DisplayTimeSpanConverter
+- HexConverter
+- SingleValueThicknessConverter
+- StringJoinConverter
+- TypeNameConverter
+- VisibilityBinding
+
 #### 2019-09-15 (v0.1.43)
-- NEW DataTemplateConverter
+- NEW DataResourceConverter
   - plugin support (+plugins for animated GIF, SVG)
-  - markup extensions (DataTemplateConverterExtension, ...)
+  - markup extensions (DataResourceConverterExtension, ...)
     - entry assembly / executing assembly / relative and absolute paths 
 
 #### 2019-09-02 (v0.1.36)